[CCUBE-2190][RYN] extract thumb & track to slider component file
